+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ## 0.7.9dev
 
+* [Fix] `psutil` is no longer a dependency for JupySQL ([#541](https://github.com/ploomber/jupysql/issues/541))
 * [Feature] Automated dependency inference when creating CTEs. `--with` is now deprecated and will display a warning. (#166)
 * [Feature] Close all connections when Python shuts down (#563)
 
 ## 0.7.8 (2023-06-01)
 
 * [Feature] Add `%sqlplot bar` and `%sqlplot pie` (#508)
+
 
 ## 0.7.7 (2023-05-31)